requirements.txt
streamlit
python-dotenv
PyPDF2
Pillow
pdf2image
openai 
